requirements.txt
Flask==2.3.3
numpy==1.24.3
opencv-python-headless==4.8.0.76
Pillow==10.0.0
scikit-learn==1.3.0
matplotlib==3.7.2
joblib==1.3.2
gunicorn==21.2.0 